LSR-881 move Asset class to Models group
